requirements.txt
configparser==5.0.2
mutagen==1.45.1
Pillow==8.2.0
tk
customtkinter